test(python): move the viewport flag regression tests to test_browser.py
